Refactor docblock parsing from Context into processor DocblockTrait
