+++ src/Context.php
@@ -189,90 +189,6 @@ class Context
     }
 
     /**
-     * A short piece of text, usually one line, providing the basic function of the associated element.
-     */
-    public function phpdocSummary(): string
-    {
-        $content = $this->phpdocContent();
-        if (!$content) {
-            return Generator::UNDEFINED;
-        }
-        $lines = preg_split('/(\n|\r\n)/', $content);
-        $summary = '';
-        foreach ($lines as $line) {
-            $summary .= $line . "\n";
-            if ($line === '' || substr($line, -1) === '.') {
-                return trim($summary);
-            }
-        }
-        $summary = trim($summary);
-        if ($summary === '') {
-            return Generator::UNDEFINED;
-        }
-
-        return $summary;
-    }
-
-    /**
-     * An optional longer piece of text providing more details on the associated element’s function. This is very
-     * useful when working with a complex element.
-     */
-    public function phpdocDescription(): string
-    {
-        $summary = $this->phpdocSummary();
-        if (!$summary) {
-            return Generator::UNDEFINED;
-        }
-
-        if (false !== ($substr = substr($this->phpdocContent(), strlen($summary)))) {
-            $description = trim($substr);
-        } else {
-            $description = '';
-        }
-        if ($description === '') {
-            return Generator::UNDEFINED;
-        }
-
-        return $description;
-    }
-
-    /**
-     * The text contents of the phpdoc comment (excl. tags).
-     */
-    public function phpdocContent(): string
-    {
-        if (Generator::isDefault($this->comment)) {
-            return Generator::UNDEFINED;
-        }
-
-        $comment = preg_split('/(\n|\r\n)/', (string) $this->comment);
-        $comment[0] = preg_replace('/[ \t]*\\/\*\*/', '', $comment[0]); // strip '/**'
-        $i = count($comment) - 1;
-        $comment[$i] = preg_replace('/\*\/[ \t]*$/', '', $comment[$i]); // strip '*/'
-        $lines = [];
-        $append = false;
-        foreach ($comment as $line) {
-            $line = ltrim($line, "\t *");
-            if (substr($line, 0, 1) === '@') {
-                break;
-            }
-            if ($append) {
-                $i = count($lines) - 1;
-                $lines[$i] = substr($lines[$i], 0, -1) . $line;
-            } else {
-                $lines[] = $line;
-            }
-            $append = (substr($line, -1) === '\\');
-        }
-        $description = trim(implode("\n", $lines));
-        if ($description === '') {
-            return Generator::UNDEFINED;
-        }
-
-        return $description;
-    }
-
-    /**
      * Create a Context based on the debug_backtrace.
      *
      * @deprecated
