+++ tests/ContextTest.php
@@ -43,77 +43,4 @@ class ContextTest extends OpenApiTestCase
         $this->assertSame('\OpenApi\Generator', $context->fullyQualifiedName('OpenApiGenerator'));
         $this->assertSame('\OpenApi\Annotations\QualifiedAlias', $context->fullyQualifiedName('OA\QualifiedAlias'));
     }
-
-    public function testPhpdocContent(): void
-    {
-        $singleLine = $this->getContext(['comment' => <<<END
-    /**
-     * A single line.
-     *
-     * @OA\Get(path="api/test1", @OA\Response(response="200", description="a response"))
-     */
-END
-        ]);
-        $this->assertEquals('A single line.', $singleLine->phpdocContent());
-
-        $multiline = $this->getContext(['comment' => <<<END
-/**
- * A description spread across
- * multiple lines.
- *
- * even blank lines
- *
- * @OA\Get(path="api/test1", @OA\Response(response="200", description="a response"))
- */
-END
-        ]);
-        $this->assertEquals("A description spread across\nmultiple lines.\n\neven blank lines", $multiline->phpdocContent());
-
-        $escapedLinebreak = $this->getContext(['comment' => <<<END
-/**
- * A single line spread across \
- * multiple lines.
- *
- * @OA\Get(path="api/test1", @OA\Response(response="200", description="a response"))
- */
-END
-        ]);
-        $this->assertEquals('A single line spread across multiple lines.', $escapedLinebreak->phpdocContent());
-    }
-
-    /**
-     * https://phpdoc.org/docs/latest/guides/docblocks.html.
-     */
-    public function testPhpdocSummaryAndDescription(): void
-    {
-        $single = $this->getContext(['comment' => '/** This is a single line DocComment. */']);
-        $this->assertEquals('This is a single line DocComment.', $single->phpdocContent());
-        $multi = $this->getContext(['comment' => "/**\n * This is a multi-line DocComment.\n */"]);
-        $this->assertEquals('This is a multi-line DocComment.', $multi->phpdocContent());
-
-        $emptyWhiteline = $this->getContext(['comment' => <<<END
-    /**
-     * This is a summary
-     *
-     * This is a description
-     */
-END
-        ]);
-        $this->assertEquals('This is a summary', $emptyWhiteline->phpdocSummary());
-        $periodNewline = $this->getContext(['comment' => <<<END
-     /**
-     * This is a summary.
-     * This is a description
-     */
-END
-        ]);
-        $this->assertEquals('This is a summary.', $periodNewline->phpdocSummary());
-        $multilineSummary = $this->getContext(['comment' => <<<END
-     /**
-     * This is a summary
-     * but this is part of the summary
-     */
-END
-        ]);
-    }
 }
